Optimisation / Geometric Corrections
First I corrected a few errors in the optimisation section that was making it so that the NTU would never show as agreeing with the optimum Q, but all it was that the other method didn't get the new m_dots and Re and when  I changed this they always agreed.
Second, I made options so that you can print the geometric values when you include display=True in the argument. I also fixed the missing tube length check, not that it's ever a restriction...